+++ src/widgets/dialogs/UserInfoPopup.hpp
@@ -6,8 +6,6 @@
 #include <pajlada/signals/scoped-connection.hpp>
 #include <pajlada/signals/signal.hpp>
 
-#include <chrono>
-
 class QCheckBox;
 
 namespace chatterino {
@@ -15,6 +13,7 @@ namespace chatterino {
 class Channel;
 using ChannelPtr = std::shared_ptr<Channel>;
 class Label;
+class EditUserNotesDialog;
 class ChannelView;
 class Split;
 
@@ -36,11 +35,13 @@ public:
 protected:
     void themeChangedEvent() override;
     void scaleChangedEvent(float scale) override;
+    void windowDeactivationEvent() override;
 
 private:
     void installEvents();
     void updateUserData();
     void updateLatestMessages();
+    void updateNotes();
 
     void loadAvatar(const QUrl &url);
     bool isMod_{};
@@ -61,6 +62,8 @@ private:
     pajlada::Signals::NoArgSignal userStateChanged_;
 
     std::unique_ptr<pajlada::Signals::ScopedConnection> refreshConnection_;
+    std::unique_ptr<pajlada::Signals::ScopedConnection>
+        userDataUpdatedConnection_;
 
     // If we should close the dialog automatically if the user clicks out
     // Set based on the "Automatically close usercard when it loses focus" setting
@@ -82,12 +85,16 @@ private:
 
         QCheckBox *block = nullptr;
         QCheckBox *ignoreHighlights = nullptr;
+        Label *notesPreview = nullptr;
+        EffectLabel2 *notesAdd = nullptr;
 
         Label *noMessagesLabel = nullptr;
         ChannelView *latestMessages = nullptr;
 
         EffectLabel2 *usercardLabel = nullptr;
     } ui_;
+
+    QPointer<EditUserNotesDialog> editUserNotesDialog_;
 
     class TimeoutWidget : public BaseWidget
     {